Migrated Properties to SelectedItemPropertiesViewModel
